implement deletion of all user clients

+++ src/AirSnitch.API/Controllers/ApiUser/ViewModels/SubscriptionPlanViewModel.cs
@@ -26,7 +26,7 @@ namespace AirSnitch.Api.Controllers.ApiUser.ViewModels
         {
             public string Period;
 
-            public ulong MaxNumberOfRequests;
+            public int MaxNumberOfRequests;
         }
     }
 }
